refactor(worklets): remove nesting of some files
