+++ packages/react-native-worklets/src/PlatformChecker.ts
