+++ packages/react-native-worklets/src/WorkletsError.ts
