+++ packages/react-native-worklets/src/errors.ts
