+++ packages/react-native-worklets/src/index.ts
@@ -1,38 +1,42 @@
 'use strict';
 
+import { initializeUIRuntime } from './initializers';
+import { WorkletsModule } from './WorkletsModule';
+
+// TODO: Specify the initialization pipeline since now there's no
+// universal source of truth for it.
+initializeUIRuntime(WorkletsModule);
+
+export type { CustomError } from './errors';
+export { createCustomError, registerCustomError } from './errors';
+export type { LoggerConfig } from './logger';
+export {
+  logger,
+  LogLevel,
+  registerLoggerConfig,
+  updateLoggerConfig,
+} from './logger';
+export { createWorkletRuntime, runOnRuntime } from './runtimes';
+export { shareableMappingCache } from './shareableMappingCache';
+export {
+  makeShareable,
+  makeShareableCloneOnUIRecursive,
+  makeShareableCloneRecursive,
+} from './shareables';
+export {
+  callMicrotasks,
+  executeOnUIRuntimeSync,
+  runOnJS,
+  runOnUI,
+  runOnUIImmediately,
+} from './threads';
+export { isWorkletFunction } from './workletFunction';
+export type { IWorkletsModule, WorkletsModuleProxy } from './WorkletsModule';
+export { WorkletsModule } from './WorkletsModule';
 export type {
-  IWorkletsModule,
-  LoggerConfig,
   ShareableRef,
   WorkletFunction,
   WorkletFunctionDev,
   WorkletRuntime,
-  WorkletsModuleProxy,
   WorkletStackDetails,
-} from './worklets';
-export {
-  callMicrotasks,
-  createCustomError,
-  createWorkletRuntime,
-  executeOnUIRuntimeSync,
-  isWorkletFunction,
-  logger,
-  LogLevel,
-  makeShareable,
-  makeShareableCloneOnUIRecursive,
-  makeShareableCloneRecursive,
-  mockedRequestAnimationFrame,
-  registerCustomError,
-  registerLoggerConfig,
-  registerWorkletStackDetails,
-  reportFatalErrorOnJS,
-  runOnJS,
-  runOnRuntime,
-  runOnUI,
-  runOnUIImmediately,
-  setupCallGuard,
-  setupConsole,
-  shareableMappingCache,
-  updateLoggerConfig,
-  WorkletsModule,
-} from './worklets';
+} from './workletTypes';
