+++ packages/react-native-worklets/src/privateGlobals.d.ts
