+++ packages/react-native-worklets/src/shareableMappingCache.ts
