+++ packages/react-native-worklets/src/threads.ts
